+++ packages/story-editor/src/components/errorBoundary/index.js
@@ -24,11 +24,11 @@ import { trackError } from '@googleforcreators/tracking';
  * Internal dependencies
  */
 import ErrorActions from './errorActions';
-import ErrorActionsWithStoryData from './errorActionsWithStoryData';
 
 class ErrorBoundary extends Component {
   static propTypes = {
     children: PropTypes.node.isRequired,
+    withStoryData: PropTypes.bool,
   };
 
   state = {
@@ -47,42 +47,16 @@ class ErrorBoundary extends Component {
   render() {
     const { error, errorInfo } = this.state;
     if (error) {
-      return <ErrorActions error={error} errorInfo={errorInfo} />;
+      return (
+        <ErrorActions
+          withStoryData={this.props.withStoryData}
+          error={error}
+          errorInfo={errorInfo}
+        />
+      );
     }
     return this.props.children;
   }
 }
 
-export class ErrorBoundaryWithStoryData extends Component {
-  static propTypes = {
-    children: PropTypes.node.isRequired,
-  };
-
-  state = {
-    error: null,
-    errorInfo: null,
-  };
-
-  componentDidCatch(error, errorInfo) {
-    this.setState({ error, errorInfo });
-    const msg = errorInfo
-      ? `${error.message}\n\n${errorInfo.componentName}\n${errorInfo.componentStack}`
-      : error.message;
-    trackError('editor_error_boundary', msg, true);
-  }
-
-  render() {
-    const { error, errorInfo } = this.state;
-    if (error) {
-      return <ErrorActionsWithStoryData error={error} errorInfo={errorInfo} />;
-    }
-    return this.props.children;
-  }
-}
-
-const shouldDisableErrorBoundaries =
-  typeof WEB_STORIES_DISABLE_ERROR_BOUNDARIES !== 'undefined' &&
-  WEB_STORIES_DISABLE_ERROR_BOUNDARIES === 'true';
-export default shouldDisableErrorBoundaries
-  ? ({ children }) => children
-  : ErrorBoundary;
+export default ErrorBoundary;
